requirements.txt
Django==2.2.6
django-cors-headers==3.1.1
django-cors-middleware==1.4.0
django-filter==2.2.0
djangorestframework==3.10.3
Jinja2==2.10.3
MarkupSafe==1.1.1
mysqlclient==1.4.4
Pillow==6.2.0
python-dateutil==2.8.0
pytz==2019.3
PyYAML==5.1.2
six==1.12.0
sqlparse==0.3.0
